FEATURE: Allow to specify `additionalFields` when subscribing a user
This change extends the `MailChimpSubscriptionFinisher` so that it
supports additional fields (aka `merge_fields`) to be sent to
MailChimp along with the email address.

Usage:

```
finishers:
  -
    identifier: 'Wwwision.Neos.MailChimp:MailChimpSubscriptionFinisher'
    options:
      listId: '<MAILCHIMP-LIST-ID>'
      additionalFields:
        'FNAME': '{firstName}'
        'LNAME': '{lastName}'
```

The `additionalFields` options expects an associative array. The values
will be processed so that `{xyz}` will be replaced by whatever the form
value for the field `xyz` was.